move iiif functions to the iiif file
