+++ common/utils/works.js
@@ -1,23 +1,10 @@
 // @flow
 import { type Work } from '../model/work';
-import {
-  type IIIFManifest,
-  type IIIFRendering,
-  type IIIFMetadata,
-  type IIIFCanvas,
-} from '../model/iiif';
+import { type IIIFRendering } from '../model/iiif';
 import { convertImageUri } from '@weco/common/utils/convert-image-uri';
 import getAugmentedLicenseInfo, {
   type LicenseData,
 } from '@weco/common/utils/licenses';
-
-export function getIIIFMetadata(
-  iiifManifest: IIIFManifest,
-  label: string
-): ?IIIFMetadata {
-  const repository = iiifManifest.metadata.find(data => data.label === label);
-  return repository;
-}
 
 export function getDigitalLocations(work: Work) {
   return work.items
@@ -31,39 +18,6 @@ export function getProductionDates(work: Work) {
   return work.production
     .map(productionEvent => productionEvent.dates.map(date => date.label))
     .reduce((a, b) => a.concat(b), []);
-}
-
-export function getDownloadOptionsFromManifest(
-  iiifManifest: IIIFManifest
-): IIIFRendering[] {
-  const sequence =
-    iiifManifest.sequences &&
-    iiifManifest.sequences.find(
-      sequence => sequence['@type'] === 'sc:Sequence'
-    );
-  const sequenceRendering = sequence && sequence.rendering;
-  const sequenceRenderingArray = Array.isArray(sequenceRendering)
-    ? sequenceRendering
-    : [sequenceRendering];
-
-  const pdfRenderingArray = iiifManifest.mediaSequences
-    ? iiifManifest.mediaSequences.reduce((acc, sequence) => {
-        return acc.concat(
-          sequence.elements
-            .map(element => {
-              return {
-                '@id': element['@id'],
-                format: element.format,
-                label: `Download ${
-                  element.format === 'application/pdf' ? 'PDF' : 'file'
-                }`,
-              };
-            })
-            .filter(Boolean)
-        );
-      }, [])
-    : [];
-  return [...sequenceRenderingArray, ...pdfRenderingArray].filter(Boolean);
 }
 
 export function getDownloadOptionsFromImageUrl(
@@ -81,82 +35,6 @@ export function getDownloadOptionsFromImageUrl(
       label: 'Download small (760px)',
     },
   ];
-}
-
-export function getCanvases(iiifManifest: IIIFManifest): IIIFCanvas[] {
-  const sequence =
-    iiifManifest.sequences &&
-    iiifManifest.sequences.find(
-      sequence =>
-        sequence['@type'] === 'sc:Sequence' &&
-        sequence.compatibilityHint !== 'displayIfContentUnsupported'
-    );
-  return sequence ? sequence.canvases : [];
-}
-
-function getManifests(iiifManifest: IIIFManifest): IIIFManifest[] {
-  return iiifManifest.manifests || null;
-}
-
-export function getManifestViewType(iiifManifest: IIIFManifest) {
-  const manifests = getManifests(iiifManifest);
-  const video =
-    iiifManifest.mediaSequences &&
-    iiifManifest.mediaSequences.find(sequence =>
-      sequence.elements.find(
-        element => element['@type'] === 'dctypes:MovingImage'
-      )
-    );
-  const audio =
-    iiifManifest.mediaSequences &&
-    iiifManifest.mediaSequences.find(sequence =>
-      sequence.elements.find(element => element['@type'] === 'dctypes:Sound')
-    );
-  const canvases = getCanvases(iiifManifest);
-  const downloadOptions = getDownloadOptionsFromManifest(iiifManifest);
-  const pdfRendering =
-    downloadOptions.find(option => option.label === 'Download PDF') || false;
-  return manifests
-    ? 'multi'
-    : audio
-    ? 'audio'
-    : video
-    ? 'video'
-    : canvases.length > 0
-    ? 'iiif'
-    : pdfRendering
-    ? 'pdf'
-    : 'none';
-}
-
-export function getVideo(iiifManifest: IIIFManifest) {
-  const videoSequence =
-    iiifManifest &&
-    iiifManifest.mediaSequences &&
-    iiifManifest.mediaSequences.find(sequence =>
-      sequence.elements.find(
-        element => element['@type'] === 'dctypes:MovingImage'
-      )
-    );
-  return (
-    videoSequence &&
-    videoSequence.elements.find(
-      element => element['@type'] === 'dctypes:MovingImage'
-    )
-  );
-}
-
-export function getAudio(iiifManifest: IIIFManifest) {
-  const videoSequence =
-    iiifManifest &&
-    iiifManifest.mediaSequences &&
-    iiifManifest.mediaSequences.find(sequence =>
-      sequence.elements.find(element => element['@type'] === 'dctypes:Sound')
-    );
-  return (
-    videoSequence &&
-    videoSequence.elements.find(element => element['@type'] === 'dctypes:Sound')
-  );
 }
 
 export function getEncoreLink(sierraId: string): string {
@@ -240,14 +118,6 @@ export function getDigitalLocationOfType(
     )
     .filter(Boolean);
   return item;
-}
-
-export function getFirstChildManifestLocation(iiifManifest: IIIFManifest) {
-  if (iiifManifest.manifests) {
-    return iiifManifest.manifests.find(manifest => manifest['@id'])['@id'];
-  } else {
-    return null;
-  }
 }
 
 type Item = Object;
